requirements-dev.txt
-r requirements.txt
bandit >= 1.7.0
black >= 20.8b1
coverage >= 5.4
flake8 >= 3.8.4
pylint >= 2.6.0
pytest >= 6.2.2
pytest-cov >= 2.11.1
requests-mock[fixture] == 1.8.0
sphinx >= 3.4.3
sphinx_rtd_theme